requirements.txt
# -----------------------------
# CORE WEB / FRAMEWORK
# -----------------------------
Django==5.2.5
Flask==2.3.2                  # Optional if using Flask utilities
gunicorn==21.2.0
whitenoise==6.9.0

# -----------------------------
# DATABASE
# -----------------------------
psycopg2-binary==2.9.10
dj-database-url==3.0.1

# -----------------------------
# SUPABASE & RELATED
# -----------------------------
supabase==1.0.3
supafunc==0.2.2
gotrue==1.3.1
postgrest==0.10.7
realtime==1.0.6

# -----------------------------
# EMAIL / SMTP
# -----------------------------
python-dotenv==1.0.1

# -----------------------------
# HTTP / REQUESTS
# -----------------------------
httpx==0.23.3
httpcore==0.16.3
requests==2.31.0
requests-toolbelt==1.0.0
h11==0.14.0
websockets==12.0
urllib3==2.5.0
anyio==4.10.0
sniffio==1.3.1

# -----------------------------
# UTILITIES
# -----------------------------
click==8.2.1
click-log==0.4.0
colorama==0.4.6
python-dateutil==2.9.0.post0
pillow==11.3.0
tqdm==4.67.1
packaging==25.0
setuptools==80.9.0
wheel==0.45.1
six==1.17.0
tzdata==2025.2
typing_extensions==4.14.1
StrEnum==0.4.15
pydantic==2.11.7
pydantic_core==2.33.2

# -----------------------------
# GIT / VERSION CONTROL
# -----------------------------
GitPython==3.1.45
gitdb==4.0.12
smmap==5.0.2
python-gitlab==3.15.0

# -----------------------------
# DOCUMENTATION / FORMATTING
# -----------------------------
docutils==0.22
readme_renderer==44.0
tomlkit==0.13.3

# -----------------------------
# SECURITY / TOKENS
# -----------------------------
deprecation==2.1.0
itsdangerous==2.2.0
keyring==25.6.0

# -----------------------------
# TEMPLATE ENGINE
# -----------------------------
Jinja2==3.1.6
MarkupSafe==3.0.2
